--- test/script/package.json
@@ -1,12 +1,0 @@
-{
-  "type": "module",
-  "dependencies": {
-    "chai": "^5.1.1",
-    "mocha": "^10.7.3",
-    "node-gyp": "^10.2.0",
-    "node-pty": "^1.0.0"
-  },
-  "scripts": {
-    "test": "mocha  --exit main.test.js"
-  }
-}
