+++ tests/script/package.json
@@ -1,15 +1,12 @@
 {
-    "type": "module",
-    "dependencies": {
-        "chai": "^5.1.1",
-        "mocha": "^10.7.3",
-        "mocha-junit-reporter": "^2.2.1",
-        "node-gyp": "^10.2.0",
-        "node-pty": "^1.0.0",
-        "winston": "^3.15.0"
-    },
-    "scripts": {
-        "test": "mocha --exit main.test.js",
-        "pre-test": "mocha --exit prerequisites.test.js"
-    }
+  "type": "module",
+  "dependencies": {
+    "chai": "^5.1.1",
+    "mocha": "^10.7.3",
+    "node-gyp": "^10.2.0",
+    "node-pty": "^1.0.0"
+  },
+  "scripts": {
+    "test": "mocha  --exit main.test.js"
+  }
 }
